--- kentik-kentikconnectpro-datasource/src/datasource/module.ts
@@ -1,9 +1,0 @@
-import { ConfigEditor } from '../components/ConfigEditor';
-import { KentikDataSource, KentikQuery } from './datasource';
-import { QueryEditor } from './QueryEditor';
-
-import { DataSourcePlugin } from '@grafana/data';
-
-export const plugin = new DataSourcePlugin<KentikDataSource, KentikQuery, {}>(KentikDataSource)
-  .setConfigEditor(ConfigEditor)
-  .setQueryEditor(QueryEditor);
